maybe fix streaming
